+++ NEWS.md
@@ -7,6 +7,8 @@
     implemented into the `vcov()` method for the output of the `estimNLR()` 
     function.
   * Wald test for implemented for the `difNLR()` function.
+  * Sandwich estimator for covariance matrix in case that `method = "nls"` was 
+    implemented into the `difNLR()` function via an argument `sandwich = TRUE`.
 
 ------